Add files via upload
Using Pandas and Seaborn to visual Fuel Price information from 2003 to 2020 to get a good feel of finding data from the UK government and understand Pandas and Seaborn better. I also learnt that Pandas does not read dates unless in a certain format for which I needed to use datetime.strptime to reformat the dates for pandas to understand them as dates.