Movie choice
Trying to decide what movie to watch on the first Sunday in lockdown. Started with a dataframe search and found that imdb has several upto date dataframes. Using the data from 2 dataframes I found the best comedy/ action movies with a average rating over 8 with the results being 'Deadpool', 'Lock, Stock...' and 'Guardians of the Galaxy'. I havent seen Deadpool in a while, seems like a good choice, bring on the pop corn. Also I have an idea in which I can use this to set up a web page for other users and help them decide movie choices based on selecting ratings/ genres/ type(movie or tv show)/ commerciality(number of ratings).