Add files via upload
Having decided to pursue Data Science and learn methodology in a new field I wanted to analyse the effect of COVID on the job market as such I did some research finding Apprenticeship information in the UK. What looks worrying but as expected the number of Apprentices taken on by recruiters in 2020 has dramatically fallen since April 2020. Also in August over the last 2 years is when the number of apprenticeships being taken on was at its highest for the year which has fallen to no new apprentices in 2020. This is a major concern as it will likely affect the apprentices which would have usually found a position not gaining the required experiences and training which will likely affect the  potential apprentices ambitions.